+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -13,5 +13,5 @@ public interface UserStorage {
 
     void deleteUser(long id);
 
-    List<User> toList();
+    List<User> getAllUserList();
 }
