+++ src/main/java/ru/yandex/practicum/filmorate/storage/database/mapper/GenreMapper.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage.database.mapper;
+
+import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.filmorate.model.Genre;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class GenreMapper implements RowMapper<Genre> {
+    @Override
+    public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return new Genre(rs.getLong("id"), rs.getString("name"));
+    }
+}
