+++ src/main/java/ru/yandex/practicum/filmorate/storage/database/mapper/RatingMapper.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage.database.mapper;
+
+import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.filmorate.model.MpaRating;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class RatingMapper implements RowMapper<MpaRating> {
+    @Override
+    public MpaRating mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return new MpaRating(rs.getLong("id"), rs.getString("name"));
+    }
+}
