+++ src/main/java/ru/yandex/practicum/filmorate/storage/database/mapper/UserMapper.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage.database.mapper;
+
+import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class UserMapper implements RowMapper<User> {
+    @Override
+    public User mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return new User(rs.getLong("id"), rs.getString("email"), rs.getString("login"), rs.getString("name"), rs.getDate("birthday").toLocalDate());
+    }
+}
